src/main/webapp/app/entities/recipes/update/recipes-update.component.ts
import { Component, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder, Validators } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize, map } from 'rxjs/operators';

import { IRecipes, Recipes } from '../recipes.model';
import { RecipesService } from '../service/recipes.service';
import { IUser } from 'app/entities/user/user.model';
import { UserService } from 'app/entities/user/user.service';

@Component({
  selector: 'jhi-recipes-update',
  templateUrl: './recipes-update.component.html',
})
export class RecipesUpdateComponent implements OnInit {
  isSaving = false;

  usersSharedCollection: IUser[] = [];

  editForm = this.fb.group({
    id: [],
    recipesName: [null, [Validators.required]],
    user: [],
  });

  constructor(
    protected recipesService: RecipesService,
    protected userService: UserService,
    protected activatedRoute: ActivatedRoute,
    protected fb: FormBuilder
  ) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ recipes }) => {
      this.updateForm(recipes);

      this.loadRelationshipsOptions();
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const recipes = this.createFromForm();
    if (recipes.id !== undefined) {
      this.subscribeToSaveResponse(this.recipesService.update(recipes));
    } else {
      this.subscribeToSaveResponse(this.recipesService.create(recipes));
    }
  }

  trackUserById(index: number, item: IUser): number {
    return item.id!;
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<IRecipes>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(recipes: IRecipes): void {
    this.editForm.patchValue({
      id: recipes.id,
      recipesName: recipes.recipesName,
      user: recipes.user,
    });

    this.usersSharedCollection = this.userService.addUserToCollectionIfMissing(this.usersSharedCollection, recipes.user);
  }

  protected loadRelationshipsOptions(): void {
    this.userService
      .query()
      .pipe(map((res: HttpResponse<IUser[]>) => res.body ?? []))
      .pipe(map((users: IUser[]) => this.userService.addUserToCollectionIfMissing(users, this.editForm.get('user')!.value)))
      .subscribe((users: IUser[]) => (this.usersSharedCollection = users));
  }

  protected createFromForm(): IRecipes {
    return {
      ...new Recipes(),
      id: this.editForm.get(['id'])!.value,
      recipesName: this.editForm.get(['recipesName'])!.value,
      user: this.editForm.get(['user'])!.value,
    };
  }
}
